+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * graph `start` property is now an array instead of a string id to support multiple entry points
 * env data structure changed from array to dictionary with array moved to `entries` property
 (e.g. `<envId>: { "label": <String>, "entries": [<entry>, ...] }`)
+* env entry `varString` property changed to `label`
 * env now has optional `label` property
 * states are now wrapped with configs automatically
 * added documentation generation